Fixed lower y value for highest x value on Discrete Uniform Graphs.
#331

The graph still shows with the lowest and the highest values cut in half, but this fix prevents the highest value from looking like its value is lower than the rest.
The added conditional on line 156 allows the other discrete functions to graph as they did before (binomial, fixed value, bernoulli, hypergeometric, poisson, and negative binomial).

Here is the explanation for the change:
Discrete Uniform graph, by default, does not display the max value. To display the users max value, a +2 modifier was added on line 101. The technical max value of the graph is still not displayed, but the users is. The +2 modifier on line 156 extends the graph wider, this was to show the actual max value of the graph, despite the fact is was never intended to.
In more simple terms, this commit changes how the graph is extended to show the actual max value.