Hiding the info overlay when in results screen.
#299

Previously the overlays themselves as well as the overlay toggle button remained visible even when in the Results section of ADSM. This commit will force hide both the overlay and the button.

The overlay retains its visibility state while the user is in the Results section, if the user ever returns to the Scenario Creator the overlay will display at its last state.