Summary Report Standard Deviation #807 - Value Update
Changed how the Standard Deviation is calculated. Previously the value was mostly calculated as part of the SQL query. Now the query simply collects the value and the python does all of the calculations to get Standard Deviation. Comments are present to make it clear exactly how the calculation is being made.

Values rounded to two places, rows that cannot have a calculated Standard Deviation are left blank.