#858 - Reverting Changes
Reverting changes relating to a secondary priority order within Destruction Priorities.

Some features are being kept, but most are being removed.

All code relating to the Secondary Priority order is simply being commented out, this is so that if a secondary priority is desired later on, it can be implemented (requires C-Engine work).