#879 - Changes for using the users unit identifier.
This commit envelops some of the changes required for ticket number 879.
A new field was added to the Unit model.
Reading values into the new field for both .csv imports and .xml imports.
Warning validation for non-unique ids.

The table on the Populations screen still needs to be switched so the new field shows instead of the user_notes field.