#893 Changing root index.html to redirect to _site/index.html
It seems the online version of the site is using index.html as the base page instead of _site/index.html. This commit changes the root index.html to redirect to _site/index.html. This will remove all formatting that might have been saved from the original gh-pages setup, but will allow it to be far more customizable in the future.